% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sendTerm.R
\name{sendTerm}
\alias{sendTerm}
\title{Create an Code field and send to terminal button shiny module}
\usage{
sendTerm(input, output, session, code, term_id, execute = function() {    
  TRUE })
}
\arguments{
\item{input}{input for the shiny session}

\item{output}{output for the shiny session}

\item{session}{session for the shiny session}

\item{code}{a reactive object for the code to be printed and executed}

\item{term_id}{RStudio terminal ID}

\item{execute}{a reactive object for whether to let the code execute after
clicking the play button.}
}
\description{
Create an Code field and send to terminal button shiny module
}
