% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sendTerm.R
\name{sendTermOutput}
\alias{sendTermOutput}
\title{Create an Code field and send to terminal button output}
\usage{
sendTermOutput(id)
}
\arguments{
\item{id}{output variable to be rendered}
}
\description{
Render a sendTerm shiny module on an application page.
}
